refactor: ThumbnailManager 인터페이스화
